code/src/requirements.txt
fastapi
uvicorn
openai
transformers
pymupdf
python-docx
mail-parser
faiss-cpu
pydantic
PyYAML
tiktoken
python-multipart
